# api/openapi.yaml
...
openapi: 3.0.3
info:
  title: Haradan API
  version: 0.1.0
  description: |-
    Phase-one Haradan HTTP API contract.

    Source of truth: docs/13 (data model) and docs/14 (use-cases).
    All paths under /api. Versioned business routes use /v1.
    BO operations require canonical role=admin AND session clientContext=ADMIN_BO.
    Client-supplied role/owner/actor values are not trusted.
    Access tokens are short-lived; refresh tokens are rotated and never stored as plaintext.
    Worker/internal use-cases are not exposed as HTTP operations.
servers:
- url: "/api"
  description: API base
...
paths:
...
  "/v1/admin/media/assets/{assetId}":
    get:
      operationId: GetAdminMediaProcessingStatus
      summary: GetAdminMediaProcessingStatus
      description: 'Exposure BO_AUTH. Actor: admin+ADMIN_BO. Requires role=admin and
        session clientContext=ADMIN_BO.'
      tags:
      - AdminMedia
      x-function-id: MEDIA-ADMIN-03
      x-exposure: BO_AUTH
      x-actor: admin+ADMIN_BO
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/MediaProcessingStatusResponse"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '500':
          "$ref": "#/components/responses/InternalError"
      security:
      - BearerAuth: []
      parameters:
      - "$ref": "#/components/parameters/AssetIdPath"
      x-error-codes:
        '400':
        - VALIDATION_ERROR
        '401':
        - UNAUTHENTICATED
        - SESSION_REVOKED
        '403':
        - FORBIDDEN
        - ACCOUNT_INACTIVE
        '404':
        - NOT_FOUND
        '500':
        - INTERNAL_ERROR
  "/v1/admin/media/assets/{assetId}/confirm":
    post:
      operationId: ConfirmAdminMediaUpload
      summary: ConfirmAdminMediaUpload
      description: 'Exposure BO_AUTH. Actor: admin+ADMIN_BO. Requires role=admin and
        session clientContext=ADMIN_BO.'
      tags:
      - AdminMedia
      x-function-id: MEDIA-ADMIN-02
      x-exposure: BO_AUTH
      x-actor: admin+ADMIN_BO
      responses:
        '202':
          description: Success
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/MediaProcessingStatusResponse"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '409':
          "$ref": "#/components/responses/Conflict"
        '500':
          "$ref": "#/components/responses/InternalError"
        '503':
          "$ref": "#/components/responses/DependencyUnavailable"
      security:
      - BearerAuth: []
      parameters:
      - "$ref": "#/components/parameters/AssetIdPath"
      x-error-codes:
        '400':
        - VALIDATION_ERROR
        '401':
        - UNAUTHENTICATED
        - SESSION_REVOKED
        '403':
        - FORBIDDEN
        - ACCOUNT_INACTIVE
        '404':
        - NOT_FOUND
        '409':
        - INVALID_STATE
        '500':
        - INTERNAL_ERROR
        '503':
        - DEPENDENCY_UNAVAILABLE
  "/v1/admin/media/uploads":
    post:
      operationId: InitiateAdminMediaUpload
      summary: InitiateAdminMediaUpload
      description: 'Exposure BO_AUTH. Actor: admin+ADMIN_BO. Requires role=admin and
        session clientContext=ADMIN_BO.'
      tags:
      - AdminMedia
      x-function-id: MEDIA-ADMIN-01
      x-exposure: BO_AUTH
      x-actor: admin+ADMIN_BO
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/InitiateMediaUploadResponse"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '422':
          "$ref": "#/components/responses/ValidationError"
        '429':
          "$ref": "#/components/responses/RateLimited"
        '500':
          "$ref": "#/components/responses/InternalError"
        '503':
          "$ref": "#/components/responses/DependencyUnavailable"
      security:
      - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/InitiateMediaUploadRequest"
      x-error-codes:
        '401':
        - UNAUTHENTICATED
        - SESSION_REVOKED
        '403':
        - FORBIDDEN
        - ACCOUNT_INACTIVE
        '422':
        - VALIDATION_ERROR
        '429':
        - RATE_LIMITED
        '500':
        - INTERNAL_ERROR
        '503':
        - DEPENDENCY_UNAVAILABLE
...
  "/v1/me/adverts/{advertId}/media":
    post:
      operationId: AttachMediaToAdvert
      summary: AttachMediaToAdvert
      description: 'Exposure FE_AUTH. Actor: ACTIVE owner. expectedMediaVersion'
      tags:
      - Media
      x-function-id: MEDIA-04
      x-exposure: FE_AUTH
      x-actor: ACTIVE owner
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/AdvertMediaCollectionResponse"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '409':
          "$ref": "#/components/responses/Conflict"
        '422':
          "$ref": "#/components/responses/ValidationError"
        '500':
          "$ref": "#/components/responses/InternalError"
      security:
      - BearerAuth: []
      parameters:
      - "$ref": "#/components/parameters/AdvertIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/AttachMediaToAdvertRequest"
      x-error-codes:
        '400':
        - VALIDATION_ERROR
        '401':
        - UNAUTHENTICATED
        - SESSION_REVOKED
        '403':
        - OWNERSHIP_REQUIRED
        - FORBIDDEN
        - ACCOUNT_INACTIVE
        '404':
        - NOT_FOUND
        '409':
        - STALE_VERSION
        - INVALID_STATE
        - CONFLICT
        '422':
        - VALIDATION_ERROR
        '500':
        - INTERNAL_ERROR
  "/v1/me/adverts/{advertId}/media/cover":
    put:
      operationId: SetAdvertCover
      summary: SetAdvertCover
      description: 'Exposure FE_AUTH. Actor: ACTIVE owner.'
      tags:
      - Media
      x-function-id: MEDIA-07
      x-exposure: FE_AUTH
      x-actor: ACTIVE owner
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/AdvertMediaCollectionResponse"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '409':
          "$ref": "#/components/responses/Conflict"
        '422':
          "$ref": "#/components/responses/ValidationError"
        '500':
          "$ref": "#/components/responses/InternalError"
      security:
      - BearerAuth: []
      parameters:
      - "$ref": "#/components/parameters/AdvertIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/SetAdvertCoverRequest"
      x-error-codes:
        '400':
        - VALIDATION_ERROR
        '401':
        - UNAUTHENTICATED
        - SESSION_REVOKED
        '403':
        - FORBIDDEN
        - ACCOUNT_INACTIVE
        '404':
        - NOT_FOUND
        '409':
        - STALE_VERSION
        - INVALID_STATE
        '422':
        - VALIDATION_ERROR
        '500':
        - INTERNAL_ERROR
  "/v1/me/adverts/{advertId}/media/order":
    put:
      operationId: ReorderAdvertMedia
      summary: ReorderAdvertMedia
      description: 'Exposure FE_AUTH. Actor: ACTIVE owner.'
      tags:
      - Media
      x-function-id: MEDIA-06
      x-exposure: FE_AUTH
      x-actor: ACTIVE owner
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/AdvertMediaCollectionResponse"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '409':
          "$ref": "#/components/responses/Conflict"
        '422':
          "$ref": "#/components/responses/ValidationError"
        '500':
          "$ref": "#/components/responses/InternalError"
      security:
      - BearerAuth: []
      parameters:
      - "$ref": "#/components/parameters/AdvertIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/ReorderAdvertMediaRequest"
      x-error-codes:
        '400':
        - VALIDATION_ERROR
        '401':
        - UNAUTHENTICATED
        - SESSION_REVOKED
        '403':
        - FORBIDDEN
        - ACCOUNT_INACTIVE
        '404':
        - NOT_FOUND
        '409':
        - STALE_VERSION
        '422':
        - VALIDATION_ERROR
        '500':
        - INTERNAL_ERROR
  "/v1/me/adverts/{advertId}/media/{assetId}":
    delete:
      operationId: DetachMediaFromAdvert
      summary: DetachMediaFromAdvert
      description: 'Exposure FE_AUTH. Actor: ACTIVE owner. expectedMediaVersion query'
      tags:
      - Media
      x-function-id: MEDIA-05
      x-exposure: FE_AUTH
      x-actor: ACTIVE owner
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/AdvertMediaCollectionResponse"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '409':
          "$ref": "#/components/responses/Conflict"
        '500':
          "$ref": "#/components/responses/InternalError"
      security:
      - BearerAuth: []
      parameters:
      - "$ref": "#/components/parameters/AdvertIdPath"
      - "$ref": "#/components/parameters/AssetIdPath"
      - "$ref": "#/components/parameters/ExpectedMediaVersionQuery"
      x-error-codes:
        '400':
        - VALIDATION_ERROR
        '401':
        - UNAUTHENTICATED
        - SESSION_REVOKED
        '403':
        - FORBIDDEN
        - ACCOUNT_INACTIVE
        '404':
        - NOT_FOUND
        '409':
        - STALE_VERSION
        - INVALID_STATE
        '500':
        - INTERNAL_ERROR
...
  "/v1/media/assets/{assetId}":
    get:
      operationId: GetMediaProcessingStatus
      summary: GetMediaProcessingStatus
      description: 'Exposure FE_AUTH. Actor: ACTIVE owner. No object keys'
      tags:
      - Media
      x-function-id: MEDIA-03
      x-exposure: FE_AUTH
      x-actor: ACTIVE owner
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/MediaProcessingStatusResponse"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '500':
          "$ref": "#/components/responses/InternalError"
      security:
      - BearerAuth: []
      parameters:
      - "$ref": "#/components/parameters/AssetIdPath"
      x-error-codes:
        '400':
        - VALIDATION_ERROR
        '401':
        - UNAUTHENTICATED
        - SESSION_REVOKED
        '403':
        - OWNERSHIP_REQUIRED
        - FORBIDDEN
        - ACCOUNT_INACTIVE
        '404':
        - NOT_FOUND
        '500':
        - INTERNAL_ERROR
  "/v1/media/assets/{assetId}/confirm":
    post:
      operationId: ConfirmMediaUpload
      summary: ConfirmMediaUpload
      description: 'Exposure FE_AUTH. Actor: ACTIVE owner. Idempotent; durable validate
        enqueued'
      tags:
      - Media
      x-function-id: MEDIA-02
      x-exposure: FE_AUTH
      x-actor: ACTIVE owner
      responses:
        '202':
          description: Success
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/MediaProcessingStatusResponse"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '409':
          "$ref": "#/components/responses/Conflict"
        '500':
          "$ref": "#/components/responses/InternalError"
        '503':
          "$ref": "#/components/responses/DependencyUnavailable"
      security:
      - BearerAuth: []
      parameters:
      - "$ref": "#/components/parameters/AssetIdPath"
      x-error-codes:
        '400':
        - VALIDATION_ERROR
        '401':
        - UNAUTHENTICATED
        - SESSION_REVOKED
        '403':
        - OWNERSHIP_REQUIRED
        - FORBIDDEN
        - ACCOUNT_INACTIVE
        '404':
        - NOT_FOUND
        '409':
        - INVALID_STATE
        - CONFLICT
        '500':
        - INTERNAL_ERROR
        '503':
        - DEPENDENCY_UNAVAILABLE
  "/v1/media/uploads":
    post:
      operationId: InitiateMediaUpload
      summary: InitiateMediaUpload
      description: 'Exposure FE_AUTH. Actor: ACTIVE user. No permanent credentials'
      tags:
      - Media
      x-function-id: MEDIA-01
      x-exposure: FE_AUTH
      x-actor: ACTIVE user
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/InitiateMediaUploadResponse"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '422':
          "$ref": "#/components/responses/ValidationError"
        '429':
          "$ref": "#/components/responses/RateLimited"
        '500':
          "$ref": "#/components/responses/InternalError"
        '503':
          "$ref": "#/components/responses/DependencyUnavailable"
      security:
      - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/InitiateMediaUploadRequest"
      x-error-codes:
        '401':
        - UNAUTHENTICATED
        - SESSION_REVOKED
        '403':
        - FORBIDDEN
        - ACCOUNT_INACTIVE
        '422':
        - VALIDATION_ERROR
        '429':
        - RATE_LIMITED
        '500':
        - INTERNAL_ERROR
        '503':
        - DEPENDENCY_UNAVAILABLE
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived access token. Session/refresh validated server-side.
        Role/owner/actor from client are not trusted.
  parameters:
...
    AdvertIdPath:
      name: advertId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    AssetIdPath:
      name: assetId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    ExpectedMediaVersionQuery:
      name: expectedMediaVersion
      in: query
      required: true
      schema:
        type: integer
        minimum: 1
...
  schemas:
...
    MediaAssetLifecycle:
      type: string
      enum:
      - UPLOAD_PENDING
      - UPLOADED
      - VALIDATING
      - MASTER_READY
      - VALIDATION_FAILED
      - CLEANUP_CANDIDATE
      - DELETING
      - PHYSICALLY_DELETED
    MediaVariantLifecycle:
      type: string
      enum:
      - PENDING
      - PROCESSING
      - READY
      - FAILED
      - DELETING
      - PHYSICALLY_DELETED
...
    ErrorDetail:
      type: object
      additionalProperties: false
      required:
      - key
      - value
      properties:
        key:
          type: string
        value:
          type: string
    FieldValidationError:
      type: object
      additionalProperties: false
      required:
      - field
      - message
      properties:
        field:
          type: string
        message:
          type: string
        code:
          type: string
    ErrorResponse:
      type: object
      required:
      - code
      - message
      - traceId
      additionalProperties: false
      properties:
        code:
          "$ref": "#/components/schemas/DomainErrorCode"
        message:
          type: string
        traceId:
          type: string
          minLength: 1
        details:
          type: array
          items:
            "$ref": "#/components/schemas/ErrorDetail"
        fieldErrors:
          type: array
          items:
            "$ref": "#/components/schemas/FieldValidationError"
...
    MediaProcessingState:
      type: object
      additionalProperties: false
      required:
      - assetId
      - lifecycleStatus
      - variants
      properties:
        assetId:
          type: string
          format: uuid
        lifecycleStatus:
          "$ref": "#/components/schemas/MediaAssetLifecycle"
        failureCode:
          type: string
          nullable: true
        failureMessage:
          type: string
          nullable: true
        variants:
          type: array
          items:
            "$ref": "#/components/schemas/MediaVariantStatusItem"
    MediaVariantStatusItem:
      type: object
      additionalProperties: false
      required:
      - transformProfile
      - lifecycleStatus
      properties:
        transformProfile:
          type: string
          minLength: 1
        lifecycleStatus:
          "$ref": "#/components/schemas/MediaVariantLifecycle"
        usage:
          type: string
          nullable: true
        publicUrl:
          type: string
          nullable: true
          description: Only when READY; never object keys
...
    OwnerMediaRelationItem:
      type: object
      additionalProperties: false
      required:
      - assetId
      - displayOrder
      - isCover
      - lifecycleStatus
      properties:
        assetId:
          type: string
          format: uuid
        displayOrder:
          type: integer
        isCover:
          type: boolean
        lifecycleStatus:
          "$ref": "#/components/schemas/MediaAssetLifecycle"
...
    InitiateMediaUploadRequest:
      type: object
      properties:
        declaredContentType:
          type: string
        declaredByteSize:
          type: integer
          minimum: 0
      additionalProperties: false
    InitiateMediaUploadResponse:
      type: object
      additionalProperties: false
      required:
      - assetId
      - upload
      - constraints
      properties:
        assetId:
          type: string
          format: uuid
        upload:
          "$ref": "#/components/schemas/UploadAuthorization"
        constraints:
          "$ref": "#/components/schemas/UploadConstraints"
    MediaProcessingStatusResponse:
      allOf:
      - "$ref": "#/components/schemas/MediaProcessingState"
    AttachMediaToAdvertRequest:
      type: object
      required:
      - assetId
      - expectedMediaVersion
      properties:
        assetId:
          type: string
          format: uuid
        displayOrder:
          type: integer
          minimum: 0
        expectedMediaVersion:
          type: integer
          minimum: 1
      additionalProperties: false
    ReorderAdvertMediaRequest:
      type: object
      required:
      - expectedMediaVersion
      - orderedAssetIds
      properties:
        expectedMediaVersion:
          type: integer
          minimum: 1
        orderedAssetIds:
          type: array
          items:
            type: string
            format: uuid
          minItems: 1
      additionalProperties: false
    SetAdvertCoverRequest:
      type: object
      required:
      - assetId
      - expectedMediaVersion
      properties:
        assetId:
          type: string
          format: uuid
        expectedMediaVersion:
          type: integer
          minimum: 1
      additionalProperties: false
    AdvertMediaCollectionResponse:
      type: object
      additionalProperties: false
      required:
      - advertId
      - mediaVersion
      - items
      properties:
        advertId:
          type: string
          format: uuid
        mediaVersion:
          type: integer
          minimum: 1
        items:
          type: array
          items:
            "$ref": "#/components/schemas/OwnerMediaRelationItem"
...
    DomainErrorCode:
      type: string
      enum:
      - VALIDATION_ERROR
      - UNAUTHENTICATED
      - FORBIDDEN
      - ACCOUNT_INACTIVE
      - EMAIL_NOT_VERIFIED
      - NOT_FOUND
      - CONFLICT
      - STALE_VERSION
      - INVALID_STATE
      - OWNERSHIP_REQUIRED
      - DUPLICATE
      - TOKEN_INVALID
      - TOKEN_EXPIRED
      - TOKEN_ALREADY_USED
      - SESSION_REVOKED
      - REFRESH_REPLAY_DETECTED
      - RATE_LIMITED
      - DEPENDENCY_UNAVAILABLE
      - PROCESSING_NOT_READY
      - QUERY_TOO_COMPLEX
      - INTERNAL_ERROR
...
    UploadAuthorization:
      type: object
      additionalProperties: false
      required:
      - method
      - url
      - expiresAt
      properties:
        method:
          type: string
          enum:
          - PUT
        url:
          type: string
          format: uri
        expiresAt:
          type: string
          format: date-time
        headers:
          type: object
          additionalProperties:
            type: string
          description: Controlled request headers for upload
      description: Provider-neutral short-lived upload authorization; no permanent
        credentials/object keys
    UploadConstraints:
      type: object
      additionalProperties: false
      required:
      - allowedContentTypes
      - maxByteSize
      - requiredHeaders
      properties:
        allowedContentTypes:
          type: array
          items:
            type: string
          minItems: 1
        maxByteSize:
          type: integer
          minimum: 1
        requiredHeaders:
          type: array
          items:
            type: string
...
  responses:
    BadRequest:
      description: Syntactic request or parameter parsing failure (malformed path/query value, UUID/integer/enum/boolean format, or unusable required parameter)
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Unauthenticated or session revoked
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Forbidden / inactive / email not verified / ownership / BO context
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/ErrorResponse"
    Conflict:
      description: Conflict / stale version / invalid state / duplicate
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/ErrorResponse"
    ValidationError:
      description: Validation or query too complex
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/ErrorResponse"
    RateLimited:
      description: Rate limited
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/ErrorResponse"
    DependencyUnavailable:
      description: External dependency unavailable
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/ErrorResponse"
    InternalError:
      description: Internal error
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/ErrorResponse"